[FEAT] 검색 기능 즐겨찾기 반환값 추가
